requirements.txt
tensorflow-cpu
pillow==9.5.0
flask
requests
flask-bootstrap
flask-wtf
keras
lxml
gunicorn
matplotlib==3.7.1
numpy==1.24.3